Cleaned special message handling
