+++ server/ServerEventType.java
@@ -4,6 +4,10 @@ public enum ServerEventType {
 
     SERVER_STARTED,
     CONNECTION_ACCEPTED,
+
+    SPECIAL_MESSAGE_RECEIVED_UNHANDLED,
+    SPECIAL_MESSAGE_RECEIVED_HANDLED,
+
     USER_JOINED,
     USER_LEFT_CHAT,
     MESSAGE_RECEIVED,
