+++ server/SpecialMessageDispatcher.java
@@ -1,20 +1,24 @@
 package concurency.chat.server;
 
 import concurency.chat.commons.CommandUtils;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.java.Log;
 
-import java.util.List;
 import java.util.Arrays;
+import java.util.List;
+import java.util.function.Consumer;
 
 import static concurency.chat.server.ServerEventType.*;
 
+@Log
+@RequiredArgsConstructor
+public class SpecialMessageDispatcher implements Consumer<ServerEvent> {
 
-public class RoomSpecialMessageHandler extends MessageHandler {
-
-    public RoomSpecialMessageHandler(EventsBus eventsBus) {
-        super(eventsBus);
-    }
+    private final EventsBus eventsBus;
 
     private static final List<String> HANDLED_COMMANDS = Arrays.asList(
+            "SEND_FILE",
+            "ACCEPT_FILE",
             "USER_JOINED_CHAT",
             "OPEN_PUBLIC_ROOM",
             "OPEN_PRIVATE_ROOM",
@@ -26,61 +30,87 @@ public class RoomSpecialMessageHandler extends MessageHandler {
     );
 
     @Override
-    public Boolean handleMessage(String message, Worker sender) {
-        Boolean wasHandled = false;
+    public void accept(ServerEvent event) {
+        switch (event.getType()) {
+            case SPECIAL_MESSAGE_RECEIVED -> dispatchMessage(event.getPayload(), event.getSource());
+        }
+    }
 
-        if (!HANDLED_COMMANDS.contains(CommandUtils.getCommandFromEvent(message)))
-            return false;
+    private void dispatchMessage(String message, Worker sender) {
 
-        String commandArgs = CommandUtils.stripCommand(message);
+        if (!HANDLED_COMMANDS.contains(CommandUtils.getCommandFromEvent(message))) {
+            eventsBus.publish(ServerEvent.builder()
+                    .type(SPECIAL_MESSAGE_RECEIVED_UNHANDLED)
+                    .payload(message)
+                    .build());
+            return;
+        }
+
+        String arguments = CommandUtils.stripCommand(message);
 
         if (message.contains("USER_JOINED_CHAT")) {
-            var name = commandArgs;
+            var name = arguments;
             sender.setName(name);
         } else if (message.contains("OPEN_PUBLIC_ROOM")) {
             eventsBus.publish(ServerEvent.builder()
                     .type(PUBLIC_ROOM_OPEN_REQUEST)
                     .source(sender)
-                    .payload(commandArgs)
+                    .payload(arguments)
                     .build());
         } else if (message.contains("OPEN_PRIVATE_ROOM")) {
             eventsBus.publish(ServerEvent.builder()
                     .type(PRIVATE_ROOM_OPEN_REQUEST)
                     .source(sender)
-                    .payload(commandArgs)
+                    .payload(arguments)
                     .build());
         } else if (message.contains("PUBLISH_TO_ROOM")) {
             eventsBus.publish(ServerEvent.builder()
                     .type(PUBLISH_TO_ROOM_REQUEST)
                     .source(sender)
-                    .payload(commandArgs)
+                    .payload(arguments)
                     .build());
         } else if (message.contains("JOIN_ROOM")) {
             eventsBus.publish(ServerEvent.builder()
                     .type(JOIN_ROOM_REQUEST)
                     .source(sender)
-                    .payload(commandArgs)
+                    .payload(arguments)
                     .build());
         } else if (message.contains("ADD_USER_TO_ROOM")) {
             eventsBus.publish(ServerEvent.builder()
                     .type(ADD_USER_TO_ROOM_REQUEST)
                     .source(sender)
-                    .payload(commandArgs)
+                    .payload(arguments)
                     .build());
         } else if (message.contains("CLOSE_ROOM")) {
             eventsBus.publish(ServerEvent.builder()
                     .type(CLOSE_ROOM_REQUEST)
                     .source(sender)
-                    .payload(commandArgs)
+                    .payload(arguments)
                     .build());
         } else if (message.contains("READ_LOG")) {
             eventsBus.publish(ServerEvent.builder()
                     .type(LOG_READ_MESSAGE)
                     .source(sender)
-                    .payload(commandArgs)
+                    .payload(arguments)
+                    .build());
+        } else if (message.contains("SEND_FILE")) {
+            eventsBus.publish(ServerEvent.builder()
+                    .type(SEND_FILE_REQUEST)
+                    .source(sender)
+                    .payload(arguments)
+                    .build());
+        } else if (message.contains("ACCEPT_FILE")) {
+            eventsBus.publish(ServerEvent.builder()
+                    .type(ACCEPT_FILE)
+                    .source(sender)
+                    .payload(arguments)
                     .build());
         }
 
-        return true;
+        eventsBus.publish(ServerEvent.builder()
+                .type(SPECIAL_MESSAGE_RECEIVED_HANDLED)
+                .payload(message)
+                .build());
+
     }
 }
